requirements.txt
discord.py>=2.6.0
python-dotenv>=1.0.0
deep-translator>=1.11.0
langdetect>=1.0.9
openai>=1.0.0
groq>=0.4.0
